+++ docs/changelog.md
@@ -7,6 +7,7 @@
 + Update API documentation for kernel updated.
 + Remake "set link length" operation.
 + Generate a QR code for mechanism expression.
++ Fix the bug of switching selection mode.
 
 **Development**: